page_rank done

+++ 04_graph/page_rank/page_rank.hpp
@@ -7,20 +7,20 @@
 
 void updatePageRank(
     std::unordered_map<uint32_t,
-        std::pair<uint16_t,
-            std::unordered_map<uint32_t, uint8_t>>>& links,
+        std::pair<uint32_t,
+            std::unordered_map<uint32_t, uint16_t>>>& links,
     const std::size_t&& loop_times = 10,
     const uint8_t&& random_rate = 5);
 
 void printAllRank(
     const std::unordered_map<uint32_t, std::string>& pages,
     std::unordered_map<uint32_t,
-        std::pair<uint16_t,
-            std::unordered_map<uint32_t, uint8_t>>>& links);
+        std::pair<uint32_t,
+            std::unordered_map<uint32_t, uint16_t>>>& links);
 
 void printTopXPages(
     const std::unordered_map<uint32_t, std::string>& pages,
     std::unordered_map<uint32_t,
-        std::pair<uint16_t,
-            std::unordered_map<uint32_t, uint8_t>>>& links,
+        std::pair<uint32_t,
+            std::unordered_map<uint32_t, uint16_t>>>& links,
     uint8_t num_of_showing_pages = 10);
